add test data for showing bad heatmap visualization
This test case shows that the heatmap also takes the number of measurements
into account, rather than just the quality of each measurement.
In a correct visualization the middle heatmap circle should be as small
as the left circle; in an incorrect visualization it is as large as the
circle on the right.